requirements.txt
# ============================================================================
# Core Web Framework
# ============================================================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
websockets==13.1

# ============================================================================
# LLM and Agent Frameworks
# ============================================================================
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langgraph==0.2.45
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# ============================================================================
# Database - MongoDB
# ============================================================================
motor==3.6.0
beanie==1.26.0
pymongo<4.10,>=4.9

# ============================================================================
# Vector Store and Embeddings
# ============================================================================
chromadb==0.5.18
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.5.1

# ============================================================================
# External APIs - PubMed/NCBI
# ============================================================================
biopython==1.84
httpx==0.27.2

# ============================================================================
# Data Validation and Settings
# ============================================================================
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# ============================================================================
# Configuration and Environment
# ============================================================================
python-dotenv==1.0.1
pyyaml==6.0.2

# ============================================================================
# Document Processing (for RAG)
# ============================================================================
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
markdown==3.7

# ============================================================================
# Utilities
# ============================================================================
tenacity==9.0.0  # Retry logic
aiofiles==24.1.0  # Async file operations
python-jose[cryptography]==3.3.0  # JWT tokens (future auth)
passlib[bcrypt]==1.7.4  # Password hashing (future auth)

# ============================================================================
# Logging and Monitoring
# ============================================================================
loguru==0.7.2
python-json-logger==3.2.1

# ============================================================================
# Development and Testing
# ============================================================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2  # For async testing
faker==33.1.0  # Generate test data

# ============================================================================
# Code Quality
# ============================================================================
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# ============================================================================
# Additional Scientific Computing (for medical calculations if needed)
# ============================================================================
numpy<2.0.0,>=1.26.0
pandas==2.2.3
